Fixed width calculation, round down
When you have a number of images in the slider which divides 100 not even, you should round down the values. Otherwise when you multiply back the values you get a number bigger than 100. So it is possible that the last image can jump to the next column. e.g (100 / 6 = 16.6667; 16.6667 * 6 = 100.002) => (100 / 6 = 16.6666; 6 * 16.6666 = 99,9996)- Smaller is better than bigger!